poc(perf): Add duplicate spans check for performance
This will be used as a commit directly on getsentry since it's experimental. Essentially this code looks for spans that match a sufficient frequency (count) to merit batching, along with hitting a cumulative duration threshold before it's worth looking into. The initial settings will be quite high just to avoid spam, but we will tune them down over the experiment, along with adding more allowed span ops, to see if we can find actual duplicate spans in our SaaS product.